Move PreferenceStorage interface to repository module.
