+++ data/local/src/main/java/jp/kuaddo/tsuidezake/data/local/internal/di/LocalDataModule.kt
@@ -2,8 +2,8 @@ package jp.kuaddo.tsuidezake.data.local.internal.di
 
 import dagger.Binds
 import dagger.Module
-import jp.kuaddo.tsuidezake.data.local.PreferenceStorage
 import jp.kuaddo.tsuidezake.data.local.internal.SharedPreferenceStorage
+import jp.kuaddo.tsuidezake.data.repository.PreferenceStorage
 
 @Module
 internal abstract class LocalDataModule {
