+++ data/repository/src/main/java/jp/kuaddo/tsuidezake/data/repository/PreferenceStorage.kt
@@ -1,4 +1,4 @@
-package jp.kuaddo.tsuidezake.data.local
+package jp.kuaddo.tsuidezake.data.repository
 
 import androidx.lifecycle.LiveData
