+++ data/repository/src/main/java/jp/kuaddo/tsuidezake/data/repository/internal/RepositoryImpl.kt
@@ -1,10 +1,10 @@
 package jp.kuaddo.tsuidezake.data.repository.internal
 
-import jp.kuaddo.tsuidezake.data.local.PreferenceStorage
 import jp.kuaddo.tsuidezake.data.remote.ApiResponse
 import jp.kuaddo.tsuidezake.data.remote.ErrorResponse
 import jp.kuaddo.tsuidezake.data.remote.SuccessResponse
 import jp.kuaddo.tsuidezake.data.remote.TsuidezakeService
+import jp.kuaddo.tsuidezake.data.repository.PreferenceStorage
 import jp.kuaddo.tsuidezake.domain.Repository
 import jp.kuaddo.tsuidezake.model.ErrorResource
 import jp.kuaddo.tsuidezake.model.Ranking
